Create DbElement class
The new class should handle basic allocation and handle the header
information. It can serve a a base class for entities not being Arrays.
