+++ src/realm/array.hpp
@@ -47,6 +47,7 @@ Searching: The main finding function is:
 
 #include <cstdint> // unint8_t etc
 
+#include <realm/db_element.hpp>
 #include <realm/util/meta.hpp>
 #include <realm/util/assert.hpp>
 #include <realm/util/file_mapper.hpp>
@@ -100,9 +101,6 @@ inline T no0(T v)
 /// found'. It is similar in function to std::string::npos.
 const size_t npos = size_t(-1);
 
-// Maximum number of bytes that the payload of an array can be
-const size_t max_array_payload = 0x00ffffffL;
-
 /// Alias for realm::npos.
 const size_t not_found = npos;
 
@@ -151,22 +149,6 @@ private:
 };
 
 
-class ArrayParent {
-public:
-    virtual ~ArrayParent() noexcept
-    {
-    }
-
-protected:
-    virtual void update_child_ref(size_t child_ndx, ref_type new_ref) = 0;
-
-    virtual ref_type get_child_ref(size_t child_ndx) const noexcept = 0;
-
-    // Used only by Array::to_dot().
-    virtual std::pair<ref_type, size_t> get_to_dot_parent(size_t ndx_in_parent) const = 0;
-
-    friend class Array;
-};
 
 struct TreeInsertBase {
     size_t m_split_offset;
@@ -221,34 +203,17 @@ struct TreeInsertBase {
 /// have a parent. Even if the original was part of a database, the copy will be
 /// free-standing, that is, not be part of any database. For intra, or inter
 /// database copying, one would have to also specify the target allocator.
-class Array : public ArrayParent {
+class Array : public DbElement, public ArrayParent {
 public:
     //    void state_init(int action, QueryState *state);
     //    bool match(int action, size_t index, int64_t value, QueryState *state);
 
     /// Create an array accessor in the unattached state.
-    explicit Array(Allocator&) noexcept;
+    using DbElement::DbElement;
 
     ~Array() noexcept override
     {
     }
-
-    enum Type {
-        type_Normal,
-
-        /// This array is the main array of an innner node of a B+-tree as used
-        /// in table columns.
-        type_InnerBptreeNode,
-
-        /// This array may contain refs to subarrays. An element whose least
-        /// significant bit is zero, is a ref pointing to a subarray. An element
-        /// whose least significant bit is one, is just a value. It is the
-        /// responsibility of the application to ensure that non-ref values have
-        /// their least significant bit set. This will generally be done by
-        /// shifting the desired vlue to the left by one bit position, and then
-        /// setting the vacated bit to one.
-        type_HasRefs
-    };
 
     /// Create a new integer array of the specified type and size, and filled
     /// with the specified value, and attach this accessor to it. This does not
@@ -258,31 +223,9 @@ public:
     /// node. It is not owned by the accessor.
     void create(Type, bool context_flag = false, size_t size = 0, int_fast64_t value = 0);
 
-    /// Reinitialize this array accessor to point to the specified new
-    /// underlying memory. This does not modify the parent reference information
-    /// of this accessor.
-    void init_from_ref(ref_type) noexcept;
-
     /// Same as init_from_ref(ref_type) but avoid the mapping of 'ref' to memory
     /// pointer.
-    void init_from_mem(MemRef) noexcept;
-
-    /// Same as `init_from_ref(get_ref_from_parent())`.
-    void init_from_parent() noexcept;
-
-    /// Update the parents reference to this child. This requires, of course,
-    /// that the parent information stored in this child is up to date. If the
-    /// parent pointer is set to null, this function has no effect.
-    void update_parent();
-
-    /// Called in the context of Group::commit() to ensure that attached
-    /// accessors stay valid across a commit. Please note that this works only
-    /// for non-transactional commits. Accessors obtained during a transaction
-    /// are always detached when the transaction ends.
-    ///
-    /// Returns true if, and only if the array has changed. If the array has not
-    /// changed, then its children are guaranteed to also not have changed.
-    bool update_from_parent(size_t old_baseline) noexcept;
+    void init_from_mem(MemRef) noexcept override;
 
     /// Change the type of an already attached array node.
     ///
@@ -312,36 +255,6 @@ public:
     /// Construct a deep copy of the specified slice of this array using the
     /// specified target allocator. Subarrays will be cloned.
     MemRef slice_and_clone_children(size_t offset, size_t slice_size, Allocator& target_alloc) const;
-
-    // Parent tracking
-    bool has_parent() const noexcept;
-    ArrayParent* get_parent() const noexcept;
-
-    /// Setting a new parent affects ownership of the attached array node, if
-    /// any. If a non-null parent is specified, and there was no parent
-    /// originally, then the caller passes ownership to the parent, and vice
-    /// versa. This assumes, of course, that the change in parentship reflects a
-    /// corresponding change in the list of children in the affected parents.
-    void set_parent(ArrayParent* parent, size_t ndx_in_parent) noexcept;
-
-    size_t get_ndx_in_parent() const noexcept;
-    void set_ndx_in_parent(size_t) noexcept;
-    void adjust_ndx_in_parent(int diff) noexcept;
-
-    /// Get the ref of this array as known to the parent. The caller must ensure
-    /// that the parent information ('pointer to parent' and 'index in parent')
-    /// is correct before calling this function.
-    ref_type get_ref_from_parent() const noexcept;
-
-    bool is_attached() const noexcept;
-
-    /// Detach from the underlying array node. This method has no effect if the
-    /// accessor is currently unattached (idempotency).
-    void detach() noexcept;
-
-    size_t size() const noexcept;
-    bool is_empty() const noexcept;
-    Type get_type() const noexcept;
 
     static void add_to_column(IntegerColumn* column, int64_t value);
 
@@ -556,28 +469,11 @@ public:
 
     bool minimum(int64_t& result, size_t start = 0, size_t end = size_t(-1), size_t* return_ndx = nullptr) const;
 
-    /// This information is guaranteed to be cached in the array accessor.
-    bool is_inner_bptree_node() const noexcept;
-
-    /// Returns true if type is either type_HasRefs or type_InnerColumnNode.
-    ///
-    /// This information is guaranteed to be cached in the array accessor.
-    bool has_refs() const noexcept;
-
-    /// This information is guaranteed to be cached in the array accessor.
-    ///
-    /// Columns and indexes can use the context bit to differentiate leaf types.
-    bool get_context_flag() const noexcept;
-    void set_context_flag(bool) noexcept;
-
-    ref_type get_ref() const noexcept;
-    MemRef get_mem() const noexcept;
-
     /// Destroy only the array that this accessor is attached to, not the
     /// children of that array. See non-static destroy_deep() for an
     /// alternative. If this accessor is already in the detached state, this
     /// function has no effect (idempotency).
-    void destroy() noexcept;
+    using DbElement::destroy;
 
     /// Recursively destroy children (as if calling
     /// clear_and_destroy_children()), then put this accessor into the detached
@@ -601,11 +497,6 @@ public:
     /// This is done by freeing the specified array node after calling
     /// destroy_deep() for every contained 'ref' element.
     static void destroy_deep(MemRef, Allocator&) noexcept;
-
-    Allocator& get_alloc() const noexcept
-    {
-        return m_alloc;
-    }
 
     // Serialization
 
@@ -805,41 +696,6 @@ public:
 
     static void get_three(const char* data, size_t ndx, ref_type& v0, ref_type& v1, ref_type& v2) noexcept;
 
-    /// The meaning of 'width' depends on the context in which this
-    /// array is used.
-    size_t get_width() const noexcept
-    {
-        return m_width;
-    }
-
-    static char* get_data_from_header(char*) noexcept;
-    static char* get_header_from_data(char*) noexcept;
-    static const char* get_data_from_header(const char*) noexcept;
-
-    enum WidthType {
-        wtype_Bits = 0,
-        wtype_Multiply = 1,
-        wtype_Ignore = 2,
-    };
-
-    static bool get_is_inner_bptree_node_from_header(const char*) noexcept;
-    static bool get_hasrefs_from_header(const char*) noexcept;
-    static bool get_context_flag_from_header(const char*) noexcept;
-    static WidthType get_wtype_from_header(const char*) noexcept;
-    static uint_least8_t get_width_from_header(const char*) noexcept;
-    static size_t get_size_from_header(const char*) noexcept;
-
-    static Type get_type_from_header(const char*) noexcept;
-
-    /// Get the number of bytes currently in use by this array. This
-    /// includes the array header, but it does not include allocated
-    /// bytes corresponding to excess capacity. The result is
-    /// guaranteed to be a multiple of 8 (i.e., 64-bit aligned).
-    ///
-    /// This number is exactly the number of bytes that will be
-    /// written by a non-recursive invocation of write().
-    size_t get_byte_size() const noexcept;
-
     /// Get the maximum number of bytes that can be written by a
     /// non-recursive invocation of write() on an array with the
     /// specified number of elements, that is, the maximum value that
@@ -871,13 +727,7 @@ public:
         }
     };
     void bptree_to_dot(std::ostream&, ToDotHandler&) const;
-    void to_dot_parent_edge(std::ostream&) const;
 #endif
-
-    static const int header_size = 8; // Number of bytes used by header
-
-    // The encryption layer relies on headers always fitting within a single page.
-    static_assert(header_size == 8, "Header must always fit in entirely on a page");
 
 private:
     Array& operator=(const Array&); // not allowed
@@ -893,40 +743,6 @@ protected:
                         ColumnBase* column) const;
 
 protected:
-    // Includes array header. Not necessarily 8-byte aligned.
-    virtual size_t calc_byte_len(size_t num_items, size_t width) const;
-
-    virtual size_t calc_item_count(size_t bytes, size_t width) const noexcept;
-
-    bool get_is_inner_bptree_node_from_header() const noexcept;
-    bool get_hasrefs_from_header() const noexcept;
-    bool get_context_flag_from_header() const noexcept;
-    WidthType get_wtype_from_header() const noexcept;
-    uint_least8_t get_width_from_header() const noexcept;
-    size_t get_size_from_header() const noexcept;
-
-    // Undefined behavior if m_alloc.is_read_only(m_ref) returns true
-    size_t get_capacity_from_header() const noexcept;
-
-    void set_header_is_inner_bptree_node(bool value) noexcept;
-    void set_header_hasrefs(bool value) noexcept;
-    void set_header_context_flag(bool value) noexcept;
-    void set_header_wtype(WidthType value) noexcept;
-    void set_header_width(int value) noexcept;
-    void set_header_size(size_t value) noexcept;
-    void set_header_capacity(size_t value) noexcept;
-
-    static void set_header_is_inner_bptree_node(bool value, char* header) noexcept;
-    static void set_header_hasrefs(bool value, char* header) noexcept;
-    static void set_header_context_flag(bool value, char* header) noexcept;
-    static void set_header_wtype(WidthType value, char* header) noexcept;
-    static void set_header_width(int value, char* header) noexcept;
-    static void set_header_size(size_t value, char* header) noexcept;
-    static void set_header_capacity(size_t value, char* header) noexcept;
-
-    static void init_header(char* header, bool is_inner_bptree_node, bool has_refs, bool context_flag,
-                            WidthType width_type, int width, size_t size, size_t capacity) noexcept;
-
 
     // This returns the minimum value ("lower bound") of the representable values
     // for the given bit width. Valid widths are 0, 1, 2, 4, 8, 16, 32, and 64.
@@ -945,8 +761,6 @@ protected:
     template <size_t width>
     void set_width() noexcept;
     void set_width(size_t) noexcept;
-    void alloc(size_t init_size, size_t width);
-    void copy_on_write();
 
 private:
     template <size_t w>
@@ -972,15 +786,6 @@ protected:
     static MemRef create(Type, bool context_flag, WidthType, size_t size, int_fast64_t value, Allocator&);
 
     static MemRef clone(MemRef header, Allocator& alloc, Allocator& target_alloc);
-
-    /// Get the address of the header of this array.
-    char* get_header() noexcept;
-
-    /// Same as get_byte_size().
-    static size_t get_byte_size_from_header(const char*) noexcept;
-
-    // Undefined behavior if array is in immutable memory
-    static size_t get_capacity_from_header(const char*) noexcept;
 
     // Overriding method in ArrayParent
     void update_child_ref(size_t, ref_type) override;
@@ -1022,10 +827,6 @@ private:
     Getter m_getter = nullptr; // cached to avoid indirection
     const VTable* m_vtable = nullptr;
 
-public:
-    // FIXME: Should not be public
-    char* m_data = nullptr; // Points to first byte after header
-
 #if REALM_ENABLE_MEMDEBUG
     // If m_no_relocation is false, then copy_on_write() will always relocate this array, regardless if it's
     // required or not. If it's true, then it will never relocate, which is currently only expeted inside
@@ -1037,26 +838,16 @@ protected:
     int64_t m_lbound; // min number that can be stored with current m_width
     int64_t m_ubound; // max number that can be stored with current m_width
 
-    size_t m_size = 0;     // Number of elements currently stored.
-    size_t m_capacity = 0; // Number of elements that fit inside the allocated memory.
-
-    Allocator& m_alloc;
-
-private:
-    size_t m_ref;
-    ArrayParent* m_parent = nullptr;
-    size_t m_ndx_in_parent = 0; // Ignored if m_parent is null.
-
-protected:
-    uint_least8_t m_width = 0;   // Size of an element (meaning depend on type of array).
-    bool m_is_inner_bptree_node; // This array is an inner node of B+-tree.
-    bool m_has_refs;             // Elements whose first bit is zero are refs to subarrays.
-    bool m_context_flag;         // Meaning depends on context.
-
 private:
     ref_type do_write_shallow(_impl::ArrayWriterBase&) const;
     ref_type do_write_deep(_impl::ArrayWriterBase&, bool only_if_modified) const;
-    static size_t calc_byte_size(WidthType wtype, size_t size, uint_least8_t width) noexcept;
+
+    // Undefined behavior if m_alloc.is_read_only(m_ref) returns true
+    size_t get_capacity_from_hdr() const noexcept
+    {
+        return DbElement::get_capacity_from_header(get_header_from_data(m_data));
+    }
+
 
     friend class SlabAlloc;
     friend class GroupWriter;
@@ -1305,42 +1096,10 @@ inline RefOrTagged::RefOrTagged(int_fast64_t value) noexcept
 {
 }
 
-inline Array::Array(Allocator& allocator) noexcept
-    : m_alloc(allocator)
-{
-}
-
 inline void Array::create(Type type, bool context_flag, size_t length, int_fast64_t value)
 {
-    MemRef mem = create_array(type, context_flag, length, value, m_alloc); // Throws
+    MemRef mem = create_array(type, context_flag, length, value, get_alloc()); // Throws
     init_from_mem(mem);
-}
-
-
-inline void Array::init_from_ref(ref_type ref) noexcept
-{
-    REALM_ASSERT_DEBUG(ref);
-    char* header = m_alloc.translate(ref);
-    init_from_mem(MemRef(header, ref, m_alloc));
-}
-
-
-inline void Array::init_from_parent() noexcept
-{
-    ref_type ref = get_ref_from_parent();
-    init_from_ref(ref);
-}
-
-
-inline Array::Type Array::get_type() const noexcept
-{
-    if (m_is_inner_bptree_node) {
-        REALM_ASSERT_DEBUG(m_has_refs);
-        return type_InnerBptreeNode;
-    }
-    if (m_has_refs)
-        return type_HasRefs;
-    return type_Normal;
 }
 
 
@@ -1387,7 +1146,7 @@ inline int64_t Array::back() const noexcept
 inline ref_type Array::get_as_ref(size_t ndx) const noexcept
 {
     REALM_ASSERT_DEBUG(is_attached());
-    REALM_ASSERT_DEBUG(m_has_refs);
+    REALM_ASSERT_DEBUG(has_refs());
     int64_t v = get(ndx);
     return to_ref(v);
 }
@@ -1416,56 +1175,16 @@ inline void Array::ensure_minimum_width(RefOrTagged ref_or_tagged)
     ensure_minimum_width(ref_or_tagged.m_value); // Throws
 }
 
-inline bool Array::is_inner_bptree_node() const noexcept
-{
-    return m_is_inner_bptree_node;
-}
-
-inline bool Array::has_refs() const noexcept
-{
-    return m_has_refs;
-}
-
-inline bool Array::get_context_flag() const noexcept
-{
-    return m_context_flag;
-}
-
-inline void Array::set_context_flag(bool value) noexcept
-{
-    m_context_flag = value;
-    set_header_context_flag(value);
-}
-
-inline ref_type Array::get_ref() const noexcept
-{
-    return m_ref;
-}
-
-inline MemRef Array::get_mem() const noexcept
-{
-    return MemRef(get_header_from_data(m_data), m_ref, m_alloc);
-}
-
-inline void Array::destroy() noexcept
-{
-    if (!is_attached())
-        return;
-    char* header = get_header_from_data(m_data);
-    m_alloc.free_(m_ref, header);
-    m_data = nullptr;
-}
-
 inline void Array::destroy_deep() noexcept
 {
     if (!is_attached())
         return;
 
-    if (m_has_refs)
+    if (has_refs())
         destroy_children();
 
     char* header = get_header_from_data(m_data);
-    m_alloc.free_(m_ref, header);
+    get_alloc().free_(get_ref(), header);
     m_data = nullptr;
 }
 
@@ -1473,10 +1192,10 @@ inline ref_type Array::write(_impl::ArrayWriterBase& out, bool deep, bool only_i
 {
     REALM_ASSERT(is_attached());
 
-    if (only_if_modified && m_alloc.is_read_only(m_ref))
-        return m_ref;
+    if (only_if_modified && get_alloc().is_read_only(get_ref()))
+        return get_ref();
 
-    if (!deep || !m_has_refs)
+    if (!deep || !has_refs())
         return do_write_shallow(out); // Throws
 
     return do_write_deep(out, only_if_modified); // Throws
@@ -1490,7 +1209,7 @@ inline ref_type Array::write(ref_type ref, Allocator& alloc, _impl::ArrayWriterB
     Array array(alloc);
     array.init_from_ref(ref);
 
-    if (!array.m_has_refs)
+    if (!array.has_refs())
         return array.do_write_shallow(out); // Throws
 
     return array.do_write_deep(out, only_if_modified); // Throws
@@ -1577,286 +1296,12 @@ inline void Array::adjust(size_t begin, size_t end, int_fast64_t diff)
         adjust(i, diff); // Throws
 }
 
-
-//-------------------------------------------------
-
-inline bool Array::get_is_inner_bptree_node_from_header(const char* header) noexcept
-{
-    typedef unsigned char uchar;
-    const uchar* h = reinterpret_cast<const uchar*>(header);
-    return (int(h[4]) & 0x80) != 0;
-}
-inline bool Array::get_hasrefs_from_header(const char* header) noexcept
-{
-    typedef unsigned char uchar;
-    const uchar* h = reinterpret_cast<const uchar*>(header);
-    return (int(h[4]) & 0x40) != 0;
-}
-inline bool Array::get_context_flag_from_header(const char* header) noexcept
-{
-    typedef unsigned char uchar;
-    const uchar* h = reinterpret_cast<const uchar*>(header);
-    return (int(h[4]) & 0x20) != 0;
-}
-inline Array::WidthType Array::get_wtype_from_header(const char* header) noexcept
-{
-    typedef unsigned char uchar;
-    const uchar* h = reinterpret_cast<const uchar*>(header);
-    return WidthType((int(h[4]) & 0x18) >> 3);
-}
-inline uint_least8_t Array::get_width_from_header(const char* header) noexcept
-{
-    typedef unsigned char uchar;
-    const uchar* h = reinterpret_cast<const uchar*>(header);
-    return uint_least8_t((1 << (int(h[4]) & 0x07)) >> 1);
-}
-inline size_t Array::get_size_from_header(const char* header) noexcept
-{
-    typedef unsigned char uchar;
-    const uchar* h = reinterpret_cast<const uchar*>(header);
-    return (size_t(h[5]) << 16) + (size_t(h[6]) << 8) + h[7];
-}
-inline size_t Array::get_capacity_from_header(const char* header) noexcept
-{
-    typedef unsigned char uchar;
-    const uchar* h = reinterpret_cast<const uchar*>(header);
-    return (size_t(h[0]) << 16) + (size_t(h[1]) << 8) + h[2];
-}
-
-
-inline char* Array::get_data_from_header(char* header) noexcept
-{
-    return header + header_size;
-}
-inline char* Array::get_header_from_data(char* data) noexcept
-{
-    return data - header_size;
-}
-inline const char* Array::get_data_from_header(const char* header) noexcept
-{
-    return get_data_from_header(const_cast<char*>(header));
-}
-
-
-inline bool Array::get_is_inner_bptree_node_from_header() const noexcept
-{
-    return get_is_inner_bptree_node_from_header(get_header_from_data(m_data));
-}
-inline bool Array::get_hasrefs_from_header() const noexcept
-{
-    return get_hasrefs_from_header(get_header_from_data(m_data));
-}
-inline bool Array::get_context_flag_from_header() const noexcept
-{
-    return get_context_flag_from_header(get_header_from_data(m_data));
-}
-inline Array::WidthType Array::get_wtype_from_header() const noexcept
-{
-    return get_wtype_from_header(get_header_from_data(m_data));
-}
-inline uint_least8_t Array::get_width_from_header() const noexcept
-{
-    return get_width_from_header(get_header_from_data(m_data));
-}
-inline size_t Array::get_size_from_header() const noexcept
-{
-    return get_size_from_header(get_header_from_data(m_data));
-}
-inline size_t Array::get_capacity_from_header() const noexcept
-{
-    return get_capacity_from_header(get_header_from_data(m_data));
-}
-
-
-inline void Array::set_header_is_inner_bptree_node(bool value, char* header) noexcept
-{
-    typedef unsigned char uchar;
-    uchar* h = reinterpret_cast<uchar*>(header);
-    h[4] = uchar((int(h[4]) & ~0x80) | int(value) << 7);
-}
-
-inline void Array::set_header_hasrefs(bool value, char* header) noexcept
-{
-    typedef unsigned char uchar;
-    uchar* h = reinterpret_cast<uchar*>(header);
-    h[4] = uchar((int(h[4]) & ~0x40) | int(value) << 6);
-}
-
-inline void Array::set_header_context_flag(bool value, char* header) noexcept
-{
-    typedef unsigned char uchar;
-    uchar* h = reinterpret_cast<uchar*>(header);
-    h[4] = uchar((int(h[4]) & ~0x20) | int(value) << 5);
-}
-
-inline void Array::set_header_wtype(WidthType value, char* header) noexcept
-{
-    // Indicates how to calculate size in bytes based on width
-    // 0: bits      (width/8) * size
-    // 1: multiply  width * size
-    // 2: ignore    1 * size
-    typedef unsigned char uchar;
-    uchar* h = reinterpret_cast<uchar*>(header);
-    h[4] = uchar((int(h[4]) & ~0x18) | int(value) << 3);
-}
-
-inline void Array::set_header_width(int value, char* header) noexcept
-{
-    // Pack width in 3 bits (log2)
-    int w = 0;
-    while (value) {
-        ++w;
-        value >>= 1;
-    }
-    REALM_ASSERT_3(w, <, 8);
-
-    typedef unsigned char uchar;
-    uchar* h = reinterpret_cast<uchar*>(header);
-    h[4] = uchar((int(h[4]) & ~0x7) | w);
-}
-
-inline void Array::set_header_size(size_t value, char* header) noexcept
-{
-    REALM_ASSERT_3(value, <=, max_array_payload);
-    typedef unsigned char uchar;
-    uchar* h = reinterpret_cast<uchar*>(header);
-    h[5] = uchar((value >> 16) & 0x000000FF);
-    h[6] = uchar((value >> 8) & 0x000000FF);
-    h[7] = uchar(value & 0x000000FF);
-}
-
-// Note: There is a copy of this function is test_alloc.cpp
-inline void Array::set_header_capacity(size_t value, char* header) noexcept
-{
-    REALM_ASSERT_3(value, <=, max_array_payload);
-    typedef unsigned char uchar;
-    uchar* h = reinterpret_cast<uchar*>(header);
-    h[0] = uchar((value >> 16) & 0x000000FF);
-    h[1] = uchar((value >> 8) & 0x000000FF);
-    h[2] = uchar(value & 0x000000FF);
-}
-
-
-inline void Array::set_header_is_inner_bptree_node(bool value) noexcept
-{
-    set_header_is_inner_bptree_node(value, get_header_from_data(m_data));
-}
-inline void Array::set_header_hasrefs(bool value) noexcept
-{
-    set_header_hasrefs(value, get_header_from_data(m_data));
-}
-inline void Array::set_header_context_flag(bool value) noexcept
-{
-    set_header_context_flag(value, get_header_from_data(m_data));
-}
-inline void Array::set_header_wtype(WidthType value) noexcept
-{
-    set_header_wtype(value, get_header_from_data(m_data));
-}
-inline void Array::set_header_width(int value) noexcept
-{
-    set_header_width(value, get_header_from_data(m_data));
-}
-inline void Array::set_header_size(size_t value) noexcept
-{
-    set_header_size(value, get_header_from_data(m_data));
-}
-inline void Array::set_header_capacity(size_t value) noexcept
-{
-    set_header_capacity(value, get_header_from_data(m_data));
-}
-
-
-inline Array::Type Array::get_type_from_header(const char* header) noexcept
-{
-    if (get_is_inner_bptree_node_from_header(header))
-        return type_InnerBptreeNode;
-    if (get_hasrefs_from_header(header))
-        return type_HasRefs;
-    return type_Normal;
-}
-
-
-inline char* Array::get_header() noexcept
-{
-    return get_header_from_data(m_data);
-}
-
-inline size_t Array::calc_byte_size(WidthType wtype, size_t size, uint_least8_t width) noexcept
-{
-    size_t num_bytes = 0;
-    switch (wtype) {
-        case wtype_Bits: {
-            // Current assumption is that size is at most 2^24 and that width is at most 64.
-            // In that case the following will never overflow. (Assuming that size_t is at least 32 bits)
-            REALM_ASSERT_3(size, <, 0x1000000);
-            size_t num_bits = size * width;
-            num_bytes = (num_bits + 7) >> 3;
-            break;
-        }
-        case wtype_Multiply: {
-            num_bytes = size * width;
-            break;
-        }
-        case wtype_Ignore:
-            num_bytes = size;
-            break;
-    }
-
-    // Ensure 8-byte alignment
-    num_bytes = (num_bytes + 7) & ~size_t(7);
-
-    num_bytes += header_size;
-
-    return num_bytes;
-}
-
-inline size_t Array::get_byte_size() const noexcept
-{
-    const char* header = get_header_from_data(m_data);
-    WidthType wtype = get_wtype_from_header(header);
-    size_t num_bytes = calc_byte_size(wtype, m_size, m_width);
-
-    REALM_ASSERT_7(m_alloc.is_read_only(m_ref), ==, true, ||, num_bytes, <=, get_capacity_from_header(header));
-
-    return num_bytes;
-}
-
-
-inline size_t Array::get_byte_size_from_header(const char* header) noexcept
-{
-    size_t size = get_size_from_header(header);
-    uint_least8_t width = get_width_from_header(header);
-    WidthType wtype = get_wtype_from_header(header);
-    size_t num_bytes = calc_byte_size(wtype, size, width);
-
-    return num_bytes;
-}
-
-
-inline void Array::init_header(char* header, bool is_inner_bptree_node, bool has_refs, bool context_flag,
-                               WidthType width_type, int width, size_t size, size_t capacity) noexcept
-{
-    // Note: Since the header layout contains unallocated bit and/or
-    // bytes, it is important that we put the entire header into a
-    // well defined state initially.
-    std::fill(header, header + header_size, 0);
-    set_header_is_inner_bptree_node(is_inner_bptree_node, header);
-    set_header_hasrefs(has_refs, header);
-    set_header_context_flag(context_flag, header);
-    set_header_wtype(width_type, header);
-    set_header_width(width, header);
-    set_header_size(size, header);
-    set_header_capacity(capacity, header);
-}
-
-
 //-------------------------------------------------
 
 inline MemRef Array::clone_deep(Allocator& target_alloc) const
 {
     char* header = get_header_from_data(m_data);
-    return clone(MemRef(header, m_ref, m_alloc), m_alloc, target_alloc); // Throws
+    return clone(MemRef(header, get_ref(), get_alloc()), get_alloc(), target_alloc); // Throws
 }
 
 inline MemRef Array::create_empty_array(Type type, bool context_flag, Allocator& alloc)
@@ -1871,79 +1316,11 @@ inline MemRef Array::create_array(Type type, bool context_flag, size_t size, int
     return create(type, context_flag, wtype_Bits, size, value, alloc); // Throws
 }
 
-inline bool Array::has_parent() const noexcept
-{
-    return m_parent != nullptr;
-}
-
-inline ArrayParent* Array::get_parent() const noexcept
-{
-    return m_parent;
-}
-
-inline void Array::set_parent(ArrayParent* parent, size_t ndx_in_parent) noexcept
-{
-    m_parent = parent;
-    m_ndx_in_parent = ndx_in_parent;
-}
-
-inline size_t Array::get_ndx_in_parent() const noexcept
-{
-    return m_ndx_in_parent;
-}
-
-inline void Array::set_ndx_in_parent(size_t ndx) noexcept
-{
-    m_ndx_in_parent = ndx;
-}
-
-inline void Array::adjust_ndx_in_parent(int diff) noexcept
-{
-    // Note that `diff` is promoted to an unsigned type, and that
-    // C++03 still guarantees the expected result regardless of the
-    // sizes of `int` and `decltype(m_ndx_in_parent)`.
-    m_ndx_in_parent += diff;
-}
-
-inline ref_type Array::get_ref_from_parent() const noexcept
-{
-    ref_type ref = m_parent->get_child_ref(m_ndx_in_parent);
-    return ref;
-}
-
-inline bool Array::is_attached() const noexcept
-{
-    return m_data != nullptr;
-}
-
-inline void Array::detach() noexcept
-{
-    m_data = nullptr;
-}
-
-inline size_t Array::size() const noexcept
-{
-    REALM_ASSERT_DEBUG(is_attached());
-    return m_size;
-}
-
-inline bool Array::is_empty() const noexcept
-{
-    return size() == 0;
-}
-
 inline size_t Array::get_max_byte_size(size_t num_elems) noexcept
 {
     int max_bytes_per_elem = 8;
     return header_size + num_elems * max_bytes_per_elem;
 }
-
-inline void Array::update_parent()
-{
-    if (m_parent)
-        m_parent->update_child_ref(m_ndx_in_parent, m_ref);
-}
-
 
 inline void Array::update_child_ref(size_t child_ndx, ref_type new_ref)
 {
